Move Preferences tests into new structure
